+++ packages/sapa/index.d.ts
@@ -643,7 +643,7 @@ declare module "@elf-framework/sapa" {
    */
   export function useStoreValue<T>(
     key: string,
-    defaultValut: T
+    defaultValue?: T
   ): [T, (value: T) => void];
 
   /**