+++ packages/sapa/src/functions/registElement.js
@@ -193,25 +193,26 @@ export function setGlobalForceRender(isForceRender = false) {
 /**
  * root instance 를 모두 그린다.
  */
-export function renderRootElementInstanceList(isForce = false) {
+export function renderRootElementInstanceList() {
   getRootElementInstanceList().forEach((instance) => {
     // root 를 재시작한다는 이야기는 최초 start 에서 생성된 root instance 를 다시 생성해야 한다는 이야기이다.
     // 그렇기 때문에 start 함수는 항상 root instance 를 생성해야 한다.
 
     const rootInstance = instance.getRootInstance();
 
-    const childInstance = rootInstance.child;
+    // const childInstance = rootInstance.child;
 
     // FIXME: 다시 연결 해주는게 맞는건가?
     // FIXME: COMPONENT_INSTANCE 의 용법을 다시 정의해야할 듯
     // FIXME: hydrate 함수 참조, 같은 방식으로 정의 되어있음.
-    if (childInstance?.$el) {
-      childInstance.$el.el[COMPONENT_INSTANCE] = childInstance;
-    }
+    // if (childInstance?.$el) {
+    //   childInstance.$el.el[COMPONENT_INSTANCE] = childInstance;
+    // }
 
     // rootInstance 의 Component 가 변겨되었는지를 검사한다.
 
-    const componentInstanceForRootRendering = childInstance || rootInstance;
+    const componentInstanceForRootRendering = rootInstance;
+    rootInstance.$el.el[COMPONENT_INSTANCE] = rootInstance;
     // rootInstance 다시 렌더링 시작
     // 기존 Hook 도 유지를 한다.
     // 그래야 다시 렌더링이 되는데, Hook 이 다시 초기화 되지 않는다.