+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @elf-framework/design-system
 
+## 0.0.220
+
+### Patch Changes
+
+- add setRefs for reload component
+
+- npm run version
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.220
+
+## 0.0.219
+
+### Patch Changes
+
+- fixed showValue property css in slider
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.219
+
 ## 0.0.216
 
 ### Patch Changes