+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @elf-framework/design-tokens
 
+## 0.0.220
+
+### Patch Changes
+
+- add setRefs for reload component
+
+- npm run version
+
+## 0.0.219
+
+### Patch Changes
+
+- fixed showValue property css in slider
+
 ## 0.0.216
 
 ### Patch Changes